Add in hasPostTypeChannel. Move createDates to shared
